добавил генерацию компонента и component.view

+++ src/generate/generateTemplate.ts
@@ -4,15 +4,19 @@ var fs = require('fs');
 
 export const generateTemplate = (dest: string, template: string, variables: IVariables) => {
    console.log('generate template to:', dest);
-   let templ = template;
+   if (!fs.existsSync(dest)) {
+      let templ = template;
 
-   templ = templ.replace(/{{ComponentName}}/g, variables.ComponentName);
+      templ = templ.replace(/{{ComponentName}}/g, variables.ComponentName);
+      variables.Ws && (templ = templ.replace(/{{Ws}}/g, variables.Ws));
+      variables.Layout && (templ = templ.replace(/{{Layout}}/g, variables.Layout));
 
-   console.log(templ);
+      console.log(templ);
 
-   try {
-      fs.writeFileSync(dest, templ);
-   } catch (e) {
-      console.error('Cannot write template to folder', dest, template, variables, e);
+      try {
+         fs.writeFileSync(dest, templ);
+      } catch (e) {
+         console.error('Cannot write template to folder', dest, template, variables, e);
+      }
    }
 };
